Rule 9 - Faster decision making (#10)
* Rule 9 - Faster decision making

Making it easier for the game to move faster. 

New text:
**203.** Voting for a rule-change is completed when one of these two conditions is met (if both conditions are true, then only the condition that happened first):
(1) The rule-change is approved by majority of players, rounding up (i.e. 3 players in a 5 player game, 2 players in a 4 player game). If this condition is met, then the rule change is adopted.
(2) The rule-change is rejected by majority of players, rounding up. If this condition is met, then the rule change is not adopted. 


Original text:
203. A rule-change is adopted if and only if the vote is unanimous among the eligible voters. If this rule is not amended by the end of the second complete circuit of turns, it automatically changes to require only a simple majority.

* Update README.md

* Update README.md

* Update README.md