Update README.md (#12)
The idea behind this rule is to add a ticking clock to the game, as well as encouraging speedy discussion and voting. It can also be used for tactical fillibustering, modeling actual politics. Because rules now only need a simple majority to pass or fail, it will not be possible for one person to misuse this rule by stalling indefinitely to secure his or her victory.